+++ tests/protocols/gmpv8/test_v7_from_gmpv8.py
@@ -572,12 +572,6 @@ class Gmpv8ModifyPortListTestCase(GmpModifyPortListTestCase, Gmpv8TestCase):
     pass
 
 
-class Gmpv8ModifyReportFormatTestCase(
-    GmpModifyReportFormatTestCase, Gmpv8TestCase
-):
-    pass
-
-
 class Gmpv8ModifyRoleTestCase(GmpModifyRoleTestCase, Gmpv8TestCase):
     pass